Null reference

+++ Assets/Scripts/UpdateStats.cs
@@ -127,11 +127,11 @@ public class UpdateStats : MonoBehaviour {
             strikesText.color = Color.red;
         }
     }
-    public int GetNumPitches()
+    public float GetNumPitches()
     {
         int numPitches = 0;
         numPitches = strikes + hits;
-        return numPitches;
+        return (float)numPitches;
     }
 }
